requirements.txt
easyocr~=1.7.1
numpy~=1.26.4
opencv-python~=4.9.0.80
scipy~=1.13.0
pillow~=10.3.0
Flask~=3.0.3
pdf2image~=1.17.0